Use a less bright white color as default
The color #fff provides higher contrast, which can cause problems to
users who experience migraines with bright colors or dyslexic users who
might see some kind of shadows around the text because of the bright
color.

So even if it isn't a big improvement, we're slighty reducing the
contrast in order to slightly reduce this problem.